Refactor performance to try to keep consistent with previous bencher records
The `definitions` test name is now a bit misleading since no definitions are
retrieved there, but it's still where user source files are ingested.
